WIP: retreave and merge Marjorie's code + remove some comments and improve code logic for login logout with simulate login temporary.
