+++ frontend/src/components/forms/auth/FormCard.tsx
@@ -1,5 +1,3 @@
-// FormCard.tsx
-
 import {
   Card,
   CardContent,
