+++ frontend/src/pages/User.tsx
@@ -4,12 +4,10 @@ import { FormCard } from '@/components/forms';
 import { Login } from '@/components/forms/auth/Login';
 import { Signup } from '@/components/forms/auth/Signup';
 import { Logout } from '@/components/forms/auth/Logout';
-import { Button } from '@/components/ui/button';
 import { useUserStore } from '@/lib/zustand/userStore';
 
 export const User = () => {
   const [isLoginMode, setIsLoginMode] = useState(false);
-
   const toggleForm = () => setIsLoginMode((prev) => !prev);
 
   const navigate = useNavigate();
@@ -18,37 +16,14 @@ export const User = () => {
 
   useEffect(() => {
     if (isAuthenticated && window.location.pathname !== '/user') {
-      navigate('/'); // redirection vers la home si l'utilisateur est authentifié et n'est pas déjà sur la page /user
+      navigate('/');
     }
   }, [isAuthenticated, navigate]);
-
-  const login = useUserStore((state) => state.login);
-  const logout = useUserStore((state) => state.logout);
-
-  const handleFakeLogin = () => {
-    login({ email: 'test@user.com', password: 'Hello123+' });
-  };
-
-  const handleLogout = () => {
-    logout();
-    navigate('/');
-  };
 
   return (
     <>
       <h1 className="mb-8 rounded-full text-center text-3xl font-thin">
         🌱 Welcome User 🌱
-        <div className="mt-6 flex justify-center">
-          {!isAuthenticated ? (
-            <Button onClick={handleFakeLogin} variant="secondary">
-              Fake connexion
-            </Button>
-          ) : (
-            <Button onClick={handleLogout} variant="secondary">
-              Se déconnecter
-            </Button>
-          )}
-        </div>
       </h1>
 
       {!isAuthenticated && (
